Homework 6. Чтение данных из Интернета с помощью Retrofit. Отображение картинок с помощью Picasso

+++ fragments/src/main/java/com/github/pashmentov96/fragments/ViewHolderListener.java
@@ -1,5 +1,5 @@
 package com.github.pashmentov96.fragments;
 
 public interface ViewHolderListener {
-    void onPersonClicked(long id);
+    void onPersonClicked(int id);
 }
